Restructure REST API server.
