+++ bin/start-server.js
@@ -1,6 +1,6 @@
 'use strict'
 
-let server = require('./server')
+let server = require('../lib/server')
 let port = process.env.PORT || 8080
 
 server.listen(port, () => {
