+++ lib/server.js
@@ -3,9 +3,8 @@
 let express = require('express')
 let app = express()
 
-let data = ['Hello', 'my', 'REST', 'API']
-
-app.get('/api/users', (req, res) => {
+app.get('/api/welcome', (req, res) => {
+  let data = { message: 'Welcome to ARROW' }
   res.json(data)
 })
